pyproject.toml
[project]
name = "l4-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.6.1",
    "google-genai>=1.25.0",
    "mcp[cli]>=1.10.1",
    "toolbox-core>=0.3.0",
]
